requirements.txt
crewai
crewai_tools
langchain_community
python-dotenv
google-generativeai
langchain-google-genai